requirements.txt
requests>=2.28.0
websocket-client>=1.4.0
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=1.0.0
